+++ src/base/b-virtual-scroll/test/runners/functional/items.js
@@ -21,12 +21,11 @@ const
 module.exports = (page) => {
 	let
 		node,
-		container;
+		container,
+		component;
 
-	beforeEach(async () => {
-		await h.component.waitForComponent(page, '#root-component');
-
-		await page.evaluate(() => {
+	const renderComponent = async (attrs = {}) => {
+		await page.evaluate(([attrs]) => {
 			globalThis.removeCreatedComponents();
 
 			const baseAttrs = {
@@ -35,35 +34,70 @@ module.exports = (page) => {
 				chunkSize: 10,
 				request: {get: {chunkSize: 10, id: Math.random()}},
 				item: 'section',
-				itemProps: ({current}) => ({'data-index': current.i})
+				itemProps: ({current}) => ({'data-index': current.i}),
+				itemKey: (data) => data.i,
+				optionKey: (data) => data.i
 			};
 
 			const scheme = [
 				{
 					attrs: {
 						...baseAttrs,
-						id: 'target'
+						id: 'target',
+						...attrs
 					}
 				}
 			];
 
 			globalThis.renderComponents('b-virtual-scroll', scheme);
-		});
+		}, [attrs]);
 
 		node = await h.dom.waitForEl(page, '#target');
+		component = await h.component.waitForComponent(page, '#target');
 		container = await h.dom.waitForRef(node, 'container');
+	};
+
+	beforeEach(async () => {
+		await h.component.waitForComponent(page, '#root-component');
+		await page.evaluate(() => globalThis.removeCreatedComponents());
 	});
 
 	describe('b-virtual-scroll with iItems trait', () => {
 		it('renders correct item', async () => {
+			await renderComponent();
 			await h.dom.waitForEl(container, 'section');
 			expect(await container.$('section')).toBeTruthy();
 		});
 
 		it('renders item with provided props', async () => {
+			await renderComponent();
 			await h.dom.waitForEl(container, 'section');
 			const attrVal = await (await container.$('section')).evaluate((el) => el.getAttribute('data-index'));
 			expect(parseInt(attrVal, 10)).toBe(0);
+		});
+
+		it('uses the deprecated `optionKey` property', async () => {
+			await renderComponent({
+				itemKey: undefined
+			});
+
+			const optionKey1 = await component.evaluate((ctx) => ctx.getItemKey({i: 0}));
+			expect(optionKey1).toBe(0);
+
+			const optionKey2 = await component.evaluate((ctx) => ctx.getItemKey({i: 1}));
+			expect(optionKey2).toBe(1);
+		});
+
+		it('uses the `itemKey` property', async () => {
+			await renderComponent({
+				optionKey: undefined
+			});
+
+			const itemKey1 = await component.evaluate((ctx) => ctx.getItemKey({i: 0}));
+			expect(itemKey1).toBe(0);
+
+			const itemKey2 = await component.evaluate((ctx) => ctx.getItemKey({i: 1}));
+			expect(itemKey2).toBe(1);
 		});
 	});